# go/gen/http/openapi.json
{"swagger":"2.0","info":{"title":"","version":"0.0.1"},"host":"localhost:80","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/watchlist":{"get":{"tags":["watchlist"],"summary":"getWatchlist watchlist","operationId":"watchlist#getWatchlist","responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/Watchlist","required":["tickers"]}},"400":{"description":"Bad Request response.","schema":{"type":"string"}},"403":{"description":"Forbidden response.","schema":{"type":"string"}},"500":{"description":"Internal Server Error response.","schema":{"type":"string"}},"502":{"description":"Bad Gateway response.","schema":{"type":"string"}},"503":{"description":"Service Unavailable response.","schema":{"type":"string"}}},"schemes":["http"]}},"/watchlist/ticker":{"post":{"tags":["watchlist"],"summary":"addWatchlistTicker watchlist","operationId":"watchlist#addWatchlistTicker","parameters":[{"name":"AddWatchlistTickerRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/Ticker"}}],"responses":{"201":{"description":"Created response.","schema":{"$ref":"#/definitions/TickerItem","required":["ticker","ohlcv"]}},"400":{"description":"Bad Request response.","schema":{"type":"string"}},"403":{"description":"Forbidden response.","schema":{"type":"string"}},"409":{"description":"Conflict response.","schema":{"type":"string"}},"500":{"description":"Internal Server Error response.","schema":{"type":"string"}},"502":{"description":"Bad Gateway response.","schema":{"type":"string"}},"503":{"description":"Service Unavailable response.","schema":{"type":"string"}}},"schemes":["http"]}},"/watchlist/{symbol}":{"delete":{"tags":["watchlist"],"summary":"removeWatchlistTicker watchlist","operationId":"watchlist#removeWatchlistTicker","parameters":[{"name":"symbol","in":"path","required":true,"type":"string"}],"responses":{"204":{"description":"No Content response."},"400":{"description":"Bad Request response.","schema":{"type":"string"}},"403":{"description":"Forbidden response.","schema":{"type":"string"}},"404":{"description":"Not Found response.","schema":{"type":"string"}},"409":{"description":"Conflict response.","schema":{"type":"string"}},"500":{"description":"Internal Server Error response.","schema":{"type":"string"}},"502":{"description":"Bad Gateway response.","schema":{"type":"string"}},"503":{"description":"Service Unavailable response.","schema":{"type":"string"}}},"schemes":["http"]}}},"definitions":{"OHLCV":{"title":"OHLCV","type":"object","properties":{"change":{"type":"number","description":"Price Change","example":20,"format":"double"},"change_percent":{"type":"number","description":"Price Change Percent","example":13.33,"format":"double"},"close":{"type":"number","description":"Close Price","example":170,"format":"double"},"high":{"type":"number","description":"High Price","example":200,"format":"double"},"last_updated_at":{"type":"integer","description":"Last Update Timestamp (Unix)","example":1768744190,"format":"int64"},"low":{"type":"number","description":"Low Price","example":50,"format":"double"},"open":{"type":"number","description":"Open Price","example":150,"format":"double"},"volume":{"type":"number","description":"Volume","example":1000000,"format":"double"}},"example":{"change":20,"change_percent":13.33,"close":170,"high":200,"last_updated_at":1768744190,"low":50,"open":150,"volume":1000000},"required":["open","high","low","close","volume","change","change_percent","last_updated_at"]},"Ticker":{"title":"Ticker","type":"object","properties":{"exchange_mic":{"type":"string","description":"Exchange MIC Code","example":"XNAS"},"name":{"type":"string","description":"Stock Name","example":"Apple Inc"},"symbol":{"type":"string","description":"Stock Symbol","example":"AAPL"}},"example":{"exchange_mic":"XNAS","name":"Apple Inc","symbol":"AAPL"},"required":["symbol"]},"TickerItem":{"title":"TickerItem","type":"object","properties":{"ohlcv":{"$ref":"#/definitions/OHLCV"},"ticker":{"$ref":"#/definitions/Ticker"}},"example":{"ohlcv":{"change":20,"change_percent":13.33,"close":170,"high":200,"last_updated_at":1768744190,"low":50,"open":150,"volume":1000000},"ticker":{"exchange_mic":"XNAS","name":"Apple Inc","symbol":"AAPL"}},"required":["ticker","ohlcv"]},"Watchlist":{"title":"Watchlist","type":"object","properties":{"tickers":{"type":"array","items":{"$ref":"#/definitions/TickerItem"},"example":[{"ohlcv":{"change":20,"change_percent":13.33,"close":170,"high":200,"last_updated_at":1768744190,"low":50,"open":150,"volume":1000000},"ticker":{"exchange_mic":"XNAS","name":"Apple Inc","symbol":"AAPL"}},{"ohlcv":{"change":20,"change_percent":13.33,"close":170,"high":200,"last_updated_at":1768744190,"low":50,"open":150,"volume":1000000},"ticker":{"exchange_mic":"XNAS","name":"Apple Inc","symbol":"AAPL"}},{"ohlcv":{"change":20,"change_percent":13.33,"close":170,"high":200,"last_updated_at":1768744190,"low":50,"open":150,"volume":1000000},"ticker":{"exchange_mic":"XNAS","name":"Apple Inc","symbol":"AAPL"}}]}},"example":{"tickers":[{"ohlcv":{"change":20,"change_percent":13.33,"close":170,"high":200,"last_updated_at":1768744190,"low":50,"open":150,"volume":1000000},"ticker":{"exchange_mic":"XNAS","name":"Apple Inc","symbol":"AAPL"}},{"ohlcv":{"change":20,"change_percent":13.33,"close":170,"high":200,"last_updated_at":1768744190,"low":50,"open":150,"volume":1000000},"ticker":{"exchange_mic":"XNAS","name":"Apple Inc","symbol":"AAPL"}},{"ohlcv":{"change":20,"change_percent":13.33,"close":170,"high":200,"last_updated_at":1768744190,"low":50,"open":150,"volume":1000000},"ticker":{"exchange_mic":"XNAS","name":"Apple Inc","symbol":"AAPL"}}]},"required":["tickers"]}}}