# go/gen/http/watchlist/server/encode_decode.go
// Code generated by goa v3.23.4, DO NOT EDIT.
//
// watchlist HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/reidlai/ta-workspace/modules/watchlist/go/design

package server

import (
	"context"
	"errors"
	"io"
	"net/http"

	watchlist "github.com/reidlai/ta-workspace/modules/watchlist/go/gen/watchlist"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeGetWatchlistResponse returns an encoder for responses returned by the
// watchlist getWatchlist endpoint.
func EncodeGetWatchlistResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*watchlist.Watchlist)
		enc := encoder(ctx, w)
		body := NewGetWatchlistResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeGetWatchlistError returns an encoder for errors returned by the
// getWatchlist watchlist endpoint.
func EncodeGetWatchlistError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "bad_request":
			var res watchlist.BadRequest
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "database_unavailable":
			var res watchlist.DatabaseUnavailable
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusServiceUnavailable)
			return enc.Encode(body)
		case "internal_error":
			var res watchlist.InternalError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		case "permission_denied":
			var res watchlist.PermissionDenied
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusForbidden)
			return enc.Encode(body)
		case "upstream_error":
			var res watchlist.UpstreamError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadGateway)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeAddWatchlistTickerResponse returns an encoder for responses returned
// by the watchlist addWatchlistTicker endpoint.
func EncodeAddWatchlistTickerResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*watchlist.TickerItem)
		enc := encoder(ctx, w)
		body := NewAddWatchlistTickerResponseBody(res)
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeAddWatchlistTickerRequest returns a decoder for requests sent to the
// watchlist addWatchlistTicker endpoint.
func DecodeAddWatchlistTickerRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*watchlist.AddWatchlistTickerPayload, error) {
	return func(r *http.Request) (*watchlist.AddWatchlistTickerPayload, error) {
		var (
			body AddWatchlistTickerRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateAddWatchlistTickerRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewAddWatchlistTickerPayload(&body)

		return payload, nil
	}
}

// EncodeAddWatchlistTickerError returns an encoder for errors returned by the
// addWatchlistTicker watchlist endpoint.
func EncodeAddWatchlistTickerError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "bad_request":
			var res watchlist.BadRequest
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "database_unavailable":
			var res watchlist.DatabaseUnavailable
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusServiceUnavailable)
			return enc.Encode(body)
		case "internal_error":
			var res watchlist.InternalError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		case "permission_denied":
			var res watchlist.PermissionDenied
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusForbidden)
			return enc.Encode(body)
		case "ticker_already_exists":
			var res watchlist.TickerAlreadyExists
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusConflict)
			return enc.Encode(body)
		case "upstream_error":
			var res watchlist.UpstreamError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadGateway)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeRemoveWatchlistTickerResponse returns an encoder for responses
// returned by the watchlist removeWatchlistTicker endpoint.
func EncodeRemoveWatchlistTickerResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		w.WriteHeader(http.StatusNoContent)
		return nil
	}
}

// DecodeRemoveWatchlistTickerRequest returns a decoder for requests sent to
// the watchlist removeWatchlistTicker endpoint.
func DecodeRemoveWatchlistTickerRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*watchlist.RemoveWatchlistTickerPayload, error) {
	return func(r *http.Request) (*watchlist.RemoveWatchlistTickerPayload, error) {
		var (
			symbol string

			params = mux.Vars(r)
		)
		symbol = params["symbol"]
		payload := NewRemoveWatchlistTickerPayload(symbol)

		return payload, nil
	}
}

// EncodeRemoveWatchlistTickerError returns an encoder for errors returned by
// the removeWatchlistTicker watchlist endpoint.
func EncodeRemoveWatchlistTickerError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "bad_request":
			var res watchlist.BadRequest
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "database_record_locked":
			var res watchlist.DatabaseRecordLocked
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusConflict)
			return enc.Encode(body)
		case "database_unavailable":
			var res watchlist.DatabaseUnavailable
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusServiceUnavailable)
			return enc.Encode(body)
		case "internal_error":
			var res watchlist.InternalError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		case "not_found":
			var res watchlist.NotFound
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusNotFound)
			return enc.Encode(body)
		case "permission_denied":
			var res watchlist.PermissionDenied
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusForbidden)
			return enc.Encode(body)
		case "upstream_error":
			var res watchlist.UpstreamError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadGateway)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// marshalWatchlistTickerItemToTickerItemResponseBody builds a value of type
// *TickerItemResponseBody from a value of type *watchlist.TickerItem.
func marshalWatchlistTickerItemToTickerItemResponseBody(v *watchlist.TickerItem) *TickerItemResponseBody {
	res := &TickerItemResponseBody{}
	if v.Ticker != nil {
		res.Ticker = marshalWatchlistTickerToTickerResponseBody(v.Ticker)
	}
	if v.Ohlcv != nil {
		res.Ohlcv = marshalWatchlistOHLCVToOHLCVResponseBody(v.Ohlcv)
	}

	return res
}

// marshalWatchlistTickerToTickerResponseBody builds a value of type
// *TickerResponseBody from a value of type *watchlist.Ticker.
func marshalWatchlistTickerToTickerResponseBody(v *watchlist.Ticker) *TickerResponseBody {
	res := &TickerResponseBody{
		Symbol:      v.Symbol,
		Name:        v.Name,
		ExchangeMic: v.ExchangeMic,
	}

	return res
}

// marshalWatchlistOHLCVToOHLCVResponseBody builds a value of type
// *OHLCVResponseBody from a value of type *watchlist.OHLCV.
func marshalWatchlistOHLCVToOHLCVResponseBody(v *watchlist.OHLCV) *OHLCVResponseBody {
	res := &OHLCVResponseBody{
		Open:          v.Open,
		High:          v.High,
		Low:           v.Low,
		Close:         v.Close,
		Volume:        v.Volume,
		Change:        v.Change,
		ChangePercent: v.ChangePercent,
		LastUpdatedAt: v.LastUpdatedAt,
	}

	return res
}